requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.3.23
langchain-community==0.3.21
langchain-openai==0.3.12
openai==1.74.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==1.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
tiktoken>=0.7,<1